requirements.txt
torch==2.0.1
transformers==4.30.0
scikit-learn==1.1.3
numpy==1.23.5
opencv-python==4.6.0.66
Pillow==9.3.0